Second version using singleton idGen pattern
This is the second version of the robot name generator.
The first version used map variable on Robot instance to enforce
unique ids. That would not work for multiple Robots.

This version uses a singleton pattern to generate the unique ids.
The singleton pattern is a design pattern that restricts the
instantiation of a class to one object. This is useful when exactly
one object is needed to coordinate actions across the system.
The concept is sometimes generalized to systems that operate more
efficiently when only one object exists, or that restrict the
instantiation to a certain number of objects. The term comes from the
mathematical concept of a singleton.